(PUP-2613) Revert change to spec test
Prior to commit 6196e75b0, file_spec.rb asserted puppet would by
default apply the sourced file's owner, group, and mode to the file
resource.

Commit 6196e75b0 changed the default behavior of `source_permissions`
from `use` to `ignore`. This caused the test's precondition to no
longer be true.

In the same commit, the spec was modified to assert that the source
mode was no longer applied to the file, but that changed the meaning
of the test. The modified test worked when running as a regular user,
but failed when running in jenkins as LocalSystem.

When puppet manages file content, it creates a
Puppet::FileSystem::Uniquefile, and renames the file. The permissions
assigned to the group depend on the default DACL for the user.

When running as a normal user, the group is "None" and the file was
created with read permissions (0x1201bf), missing DELETE (0x10000),
WRITE_DAC (0x40000), and WRITE_OWNER (0x80000). This caused puppet
to report the mode as 0740, which didn't match 0770, so the test
passed.

When run in jenkins as LocalSystem, the group is "SYSTEM" and puppet
will always ensure that SYSTEM has full control. This caused puppet to
report the mode as 0770, which caused the test to fail.

This commit reverts the change to the spec test, and specifies the
precondition `source_permissions => use`. The test will eventually be
removed in PUP-2609, since we're removing `use` behavior on Windows.