(WIP,PUP-3500) Ignore environment when handling setting hooks
Env was only being used when handling hooks that were not defered to
after initialization.

This breaks settings_spec "Puppet::Settings when parsing its
configuration should pass the environment-specific value to the hook
when one is available"

Not sure if ths is important (and whether it is relevant only for legacy
settings or also for directory environment settings).