(PUP-3272) Change Status yaml deserialization to support psych
This changes deserialization of a Status instance to use
a custom initialization (based on params hash). The implementation
is in a PsychSupport module to make it more visible and easier to add to
other classes needed the same support.

Since this change is not compatible with 1.8.7, the report_spec test
that depends on this functionality is restricted to not run on Ruby
1.8.7.