(PUP-3272) Remove special treatment of YAML encoded parameters
When it was allowed to pass YAML on the network it was also allowed to
pass parameters that were YAML encoded. The 3x logic had deprecation
warnings in place for these places.

Now in 4x, when YAML is no longer supported, the request parameters are
simply always strings. If someone yaml encodes data and sets it in a
parameter this has no special meaning - it is just a string that happens
to contain yaml (e.g. "--- \n 'foo'\n").

This means that it is now ok to have parameters that look like YAML but
it is not - this could cause problems earlier.