(PUP-3277) Ensure that user and group are set in func4x test
Without these settings, the puppet apply will reset Puppet directories
and files as root.  This doesn't impact a foss test run, but a pe test
run using the puppetserver will fail to start up master threads unless
there is an intervening agent or apply run with a puppet.conf that has
user/group with service account settings.  The puppetserver master
threads were failing because of permissions on the ca.pem.