(maint) Remove extra agent from the debian init script
Previously when the debian puppet service was started, it would have
an extra agent in the name because we set both daemon_opts and name to
agent and pass both to the starting exec. This commit addresses that by
setting daemon_opts to be empty, it can be overridden in the defaults
file.