(PUP-3555) Add override-able factory for environment cache entries
This commit adds a new `cache_entry_factory_class` variable that
allows a caller to override the default implementation of the
logic used to create a cache entry for an environment.

It also abstracts out the method that is used to retrieve the
current system time, so that it can be passed in as a parameter
to the factory constructor to support testing.