(PUP-3277) Allow resource changes in ticket 5477 test
Reports ending up in the /var/log/pe-puppet-dashboard/failed_reports
directory during pe tests runs had an seluser context that differed from
that which the puppet_enterprise module would manage in subsequent agent
runs. This was interaction between the
acceptance/tests/allow_arbitrary_node_name_fact_for_agent.rb from
which the reports were being generated, and this ticket_5477 test which
was exiting with a 2 because of the selinux context changes in the
failed_reports directory. I'm not positive why these two particular
tests happen to have this interaction. The allow_arbitrary_node_name
test is the first to run after installation. The ticket_5477 test is not
the first test to run with an agent -t after this, but maybe the first
that ran without accepting an exit of 2. I don't see anything
wrong with the report files themselves being adjusted by the pe modules,
so I'm relaxing the acceptable_exit_codes in ticket_5477, because I have
run out of other ideas.