(PUP-3568) Use puppet-lint v1 in module skeleton
The relative bit is necessary with puppet-lint v1 or it will fail if the
directory is not named correctly, which anyone doing development is
going to have. After I upgraded to puppet-lint v1, all of my travis-ci
builds started to fail and this is the fix.

These are all common situations that would fail

* Using Travis-ci.org

* `puppet module generate ghoneycutt-foo && cd ghoneycutt-foo && \
  rake lint`

* `git clone https://github.com/ghoneycutt/puppet-module-dnsclient.git \
  && cd puppet-module-dnsclient && rake lint`