(PUP-3376) Log critical errors at the error level
Commit 1d5386a modified the windows daemon to observe puppet's
log_level. This added `log_XXX` methods for alert, emerg, and crit log
levels. Although the service never tried to log at those levels, if it
did, the message would have been logged to the Application event log as
EVENTLOG_WARNING_TYPE.

This commit future-proofs the code to ensure that :alert, :emerg, and
:crit are logged as EVENTLOG_ERROR_TYPE.