(PUP-3272) Fix assymetric serialization - deserialiation in fact.rb
The fix is to incude the Puppet::Util::PsychSupport which deseerializes
by passing the read hash on to initialize_from_hash. If this is not
done, then all values are poked directly into instance variables
without any chance to affect formatting / data type.

This changes was required because tip of master made facts.rb use
assymetrical serialization/deserialization.