(PUP-3388) Zpool mirrors
https://tickets.puppetlabs.com/browse/PUP-3388

Without this patch applied, multiple drives with mirrors could not be added. This is a problem because the resoruce specifies that it could handle these conditions.

With the 1 line of code that I removed, and the 2 lines of codes that I added I was finally able to satisfy what the resource was intended to do