(PUP-3268) Remove use of templatedir setting from parser/files
The templatedir setting is deprecated and being removed for Puppet 4.
It was used only in Puppet::Parser::Files to provide an alternate path
location for templates, instead of looking them up in modules. This
commit removes the use of templatedir, allowing paths to be looked up
only from an absolute path or relative to a module.