(maint) Refactor parser/files find helpers
Puppet::Parser::Files#find_file and #find_template methods were
performing the same work to vet an absolute path and lookup a module
prior to calling an implementation specific method on that module. The
find_template method had a helper which this commit generalizes to work
for both find_file and find_template.