(maint) Add an inspect method to node/environment
Debugging environments often requires seeing what manifest and
modulepath have been set in addition to just the name of the
environment.  Environment.to_s just prints the name.  To help with
debugging, I've filled out the inspect method to display this additional
information.  This will be used in pretty printing, or when explicitly
called.