(PUP-3560) ASN.1 encoding/decoding handled properly
Because Ruby's OpenSSL lib munges unknown extension values so bad
(it turns non-printable chars into '.' characters) there's no way to DER
decode them. I ended up having to DER decode the entire certificate and
decode the extensions manually.

Regarding test changes, due to a bug in OpenSSL::X509::Extension the netscape comment is not
parsed correctly now that it is being properly DER encoded. I ammended
the test to compare the comment string with the buggy characters in it.