(PUP-3268) Remove --manifestdir reference from application/doc
The doc application used --manifestdir as a means of specifying the
directory of manifests to generate docs for in rdoc mode.  Since
manifestdir is being removed for Puppet 4, and the Puppet Strings module
is being used instead for all manifest documentation, this commit
removes references to manifestdir in the doc application.