(PUP-2366) Use EL 7 masters for Windows acceptance
Prior to this commit, our Windows acceptance tests used RHEL 6 masters,
which ship with Ruby 1.8.7. As part of PUP-2366, we need to bump these to
EL 7 to make use of a newer Ruby. We would also periodically run into
issues where there was a transient failure on the master that caused the
Windows acceptance job to fail.

This commit modifies all the Windows Beaker node configs to use RHEL 7
masters.