(PUP-3268) Set directory environments as default
Prior to this commit, environmentpath defaulted to the empty string,
which meant that legacy environment loaders would be used. As a first
step to removing the legacy environment code, environmentpath is now set
to "$confdir/environments", defaulting to directory loaders.

In addition, this commit completely removes the deprecated manifestdir
and templatedir settings. It also removes the defaults for manifest and
modulepath, which are only to be allowed in environment.conf and on the
commandline. Enforcement of that is still to be done, as noted by the
XXX_ENV comments. Manifest's default is changed to '/dev/null'.