(PUP-3642) Remove environment name from URL path
This commit makes some changes to the indirector network layer
such that the environment name is no longer a part of the main
URL path structure.  Instead, the environment becomes a query
parameter for GET requests, or part of the form body for POST
requests.  This will allow us to have a more finite URL space
and run Puppet in the same webserver with other apps, without
worrying about URL collisions.