(PUP-3274) Make the 4x handling of relatie references standard.
When the future parser was in effect, all references to resources are
made absolute (as well as asserted to have the correct form). This
change make the 4x handling of references (for functions include,
contain, require, and hiera_include) have the 4x behavior.

As a consequence, tests that have not earlier been executed with the
setting Puppet[:parser] = 'future' have been changed to reflect the
change in behavior.