(PUP-3274) Make all spec tests run parser future by default
This changes the default parser setting to 'future' thereby making
all tests that are not explicit about future/current parser to run
with the future parser.

This is one step in the direction of making the future parser the
one and only parser.

Tests that are no longer needed (only running for parser current
and covered by other spec tests for future parser have been remvoved,
and tests that fail because of syntax and semantic changes have been
updated.

The future_compiler_spec has been renamed to compiler4_spec.