(PUP-3379) Don't retry if block raises fatal error
Previously, if the caller didn't specify which exceptions to retry, the
retry_action method would retry all exceptions, even SystemExit and
NoMemoryError, which should never be rescued.

This method modifies the retry_action method to always re-raise
SystemExit and NoMemoryError, even in the situation where the caller
asks for them to be retried.