(PUP-3568) Require at least 0.8.2 of puppetlabs_spec_helper
Previously the puppet-lint relative option was specified in the Rakefile
generated by the module tool. This option allows `rake lint` to succeed
when the module's top-level directory is of the form <author>-<name>,
which is usually the case during module development and testing. Without
the option puppet-lint will report:

    manifests/init.pp - ERROR: <name> not in autoload module layout on line 38

This commit updates the puppetlabs_spec_helper dependency to 0.8.2,
which already handles the relative option[1].

[1] https://github.com/puppetlabs/puppetlabs_spec_helper/commit/958102eb9