(PUP-3274) Fix failing tests after rebase to latest master
The rebase brought in the new collector implementation. It has changes
that mocking tests failed.

There is still one failing tests due to a real problem (PUP-3665) that
will be fixed and included in a subsequent rebase of this commit.
The failing tests is: ./spec/unit/transaction_spec.rb:649 which gets a
type system PResourceType from 4x instead of a Puppet::Resource
instance.