(PUP-3638) Remove .pp processing from RDoc, and remove RDoc1 support
This removes all processing of .pp files in the rdoc support in puppet.
Only the processing of elements expressed in .rb files are kept
(i.e. functions, types, facts).

This also drops supports for RDoc1, since support for Ruby 1.8.7 is
also dropped.

While further cleanup, removal, and tidying of tests could be performed,
all of the Rdoc support is slated for removal when the new documentation
tool is finished and also handles the functions, types and facts. (This
includes fixtures, helpers, methods that can be (probably) be removed
and that now raises exceptions or does nothing, etc.

Remove test of RDoc parsing .pp