(PUP-1035) Remove server component from pluginsource and pluginfactsource
Previously, the pluginsource and pluginfactsource settings would always
use the $server as the host component for their respective URLs. This is
fine for non-SRV environments, but is not compatible in SRV
environments.

This change will allow pluginsync and pluginfactsync to use SRV records
to determine which host to connect to. This change should not affect
environments where SRV records are not being used.