(PUP-3388) Differentiate between single and multi mirror and raidz pools
Previously, the acceptance test attempted to create a single mirrored
pool with 3 virtual devices, and a single raidz with 3 virtual devices.
The syntax does not match the type specification, which says:

    List of all the devices to mirror for this pool. Each mirror should be a
    space separated string.

And the same thing for raidz.

This commit modifies the acceptance test so that it uses valid mirror
and raidz syntax. It can either be a single pool with multiple devices,
e.g. mirror => ['a b'], or multiple pools each with multiple devices,
e.g. mirror => ['a b', 'c d'].