(PUP-3274) Update documentation and rephrase mentions of "future parser"
This updates the language in the documentation of functions that had
conditional behavior if future parser was used.

The since meta tags are updated to 4.0.0; to learn details about in
which 3x version something was introduced in future parser requires
looking at the relevant documentation for that 3x version. (We don't
want to drag the "future parser" backage along.

Comments (and documentation) rephrased from using "future parser" as 4x
(if required), or just stating "how things are".