(PUP-3341) use the proper manifest with puppet apply and ENC
The apply application works by overriding the :current_environment with the
file that was specified on the command line. This had no effect when the node
object (from the ENC) specified an environment of its own.

Remedy this by looking up the node before the environment, favoring its
environment over the configured value and overriding the :manifest of the
chosen one.