(PUP-3268) Do not lookup [legacy] puppet.conf sections
A named puppet.conf section other than main, master, agent, user was
taken to be an overriding block of environment specific settings prior
to this commit. This behavior provided manifest, modulepath,
config_version when working with a legacy environment, and would
override other settings if the code path looking up a particular setting
happened to submit environment (not all do). With legacy environment
removal, support for reading settings from environment puppet.conf
sections is also dropped. This allowed some additional cleanup of the
settings searchpath and value_sets_for code, in particular, abstracting
the type of the searchpath element so that we can cleanly separate
whether we are obtaining a puppet.conf section named 'main' or an
environment.conf for environment 'main' for instance.