(PUP-3268) Fail if environment settings encounted in puppet.conf
The manifest, modulepath, and config_version settings are environment
specific and were deprecated specifically when being set in puppet.conf,
as opposed to on the commandline or in an environment.conf file.  For
Puppet 4, we are now raising an error if these settings are found in
puppet.conf during initialization.